Move payment profile creation logic into form
