+++ authorizenet/views.py
@@ -6,7 +6,7 @@ except ImportError:
 from django.conf import settings
 from django.shortcuts import render
 from django.views.decorators.csrf import csrf_exempt
-from django.views.generic.edit import FormView
+from django.views.generic.edit import CreateView
 
 from authorizenet.forms import AIMPaymentForm, BillingAddressForm, CustomerPaymentForm
 from authorizenet.models import CustomerProfile, CustomerPaymentProfile
@@ -125,30 +125,11 @@ class AIMPayment(object):
         )
 
 
-class PaymentProfileCreationView(FormView):
+class PaymentProfileCreationView(CreateView):
     template_name = 'authorizenet/create_payment_profile.html'
     form_class = CustomerPaymentForm
 
-    def form_valid(self, form):
-        """If the form is valid, save the payment profile"""
-        data = form.cleaned_data
-        self.create_payment_profile(payment_data=data, billing_data=data)
-        return super(PaymentProfileCreationView, self).form_valid(form)
-
-    def create_payment_profile(self, **kwargs):
-        """Create and return payment profile"""
-        customer_profile = self.get_customer_profile()
-        if customer_profile:
-            return CustomerPaymentProfile.objects.create(
-                customer_profile=customer_profile, **kwargs)
-        else:
-            customer_profile = CustomerProfile.objects.create(
-                user=self.request.user, **kwargs)
-            return customer_profile.payment_profiles.get()
-
-    def get_customer_profile(self):
-        """Return customer profile or ``None`` if none exists"""
-        try:
-            return CustomerProfile.objects.get(user=self.request.user)
-        except CustomerProfile.DoesNotExist:
-            return None
+    def get_form_kwargs(self):
+        kwargs = super(PaymentProfileCreationView, self).get_form_kwargs()
+        kwargs['user'] = self.request.user
+        return kwargs
